Added a full blown Networking API wrapping the WebSocket protocol.
For SGF, all game clients must be able to at least be able to connect
to a WebSocket server. Able game clients can implement a WebSocket
server that can allow for peer-to-peer gaming.

The current Java implementation uses a WebSocket client and server
class written by me. The browser uses the native WebSocket API if
available, and falls back to a Flash emulation layer if not.